requirements_mac_final.txt
faster-whisper==1.1.1
livekit-agents==1.1.4
livekit-plugins-noise-cancellation
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-turn-detector
memoripy==0.1.2
python-dotenv
soundfile
ctranslate2==4.3.1
langchain-openai
openai
httpx
numpy
pydantic
PyYAML
